task 2 (partly) not ready: filter in URL

+++ src/Category.js
@@ -7,6 +7,7 @@ class Category extends React.Component {
     constructor(props) {
         super(props);
         props.data.selected = props.data.selected || false;
+        props.data.showChildren = true;
         this.state = props.data;
     }
 
@@ -34,35 +35,51 @@ class Category extends React.Component {
         this.props.handleUpdateList();
     }
 
+    handleCollapseExpandClick(e) {
+        console.log("Category:handleCollapseExpandClick");
+        let state = this.state;
+        state.showChildren = !this.state.showChildren;
+        state.showChildren ? e.target.innerHTML = "^" : e.target.innerHTML = ">";
+        this.setState(state);
+    }
+
     render() {
-        return (
-            <div className="Category">
-                <div className={this.state.selected ? "Category-content Category-selected" : "Category-content"}>
-                    {this.state.children ? <span className="Category-expand">^</span> :
-                        <span className="Category-expand"/>}
-                    <span className="Category-title" onClick={this.handleSelectCategory.bind(this)}>
+        if ((!this.props.filter.showDone && this.state.isCompleted)) {
+            return (null);
+        } else {
+            return (
+                <div className="Category">
+                    <div className={this.state.selected ? "Category-content Category-selected" : "Category-content"}>
+                        {this.state.children && this.state.children.length > 0 ?
+                            <span onClick={this.handleCollapseExpandClick.bind(this)}
+                                  className="Category-expand">^</span> :
+                            <span className="Category-expand"/>}
+                        <span className="Category-title" onClick={this.handleSelectCategory.bind(this)}>
                             {this.state.title}
                         </span>
-                    <button onClick={this.handleEditClick.bind(this)}>Edit</button>
-                    <button onClick={() => {
-                        if (confirm('Delete the category?')) {
-                            this.handleDeleteClick();
-                        }
-                    }}>Delete
-                    </button>
-                    <button onClick={this.handleAddCategoryClick.bind(this)}>+</button>
+                        <button onClick={this.handleEditClick.bind(this)}>Edit</button>
+                        <button onClick={() => {
+                            if (confirm('Delete the category?')) {
+                                this.handleDeleteClick();
+                            }
+                        }}>
+                            Delete
+                        </button>
+                        <button onClick={this.handleAddCategoryClick.bind(this)}>+</button>
+                    </div>
+                    {this.state.children && this.state.showChildren ?
+                        <div className="Category-sublist">
+                            {this.state.children.map((elem) => <Category
+                                handleSelectCategory={this.props.handleSelectCategory}
+                                handleUpdateList={this.props.handleUpdateList}
+                                filter={this.props.filter}
+                                key={elem.key}
+                                data={elem}/>)}
+                        </div> : ""
+                    }
                 </div>
-                {this.state.children ?
-                    <div className="Category-sublist">
-                        {this.state.children.map((elem) => <Category
-                            handleSelectCategory={this.props.handleSelectCategory}
-                            handleUpdateList={this.props.handleUpdateList}
-                            key={elem.key}
-                            data={elem}/>)}
-                    </div> : ""
-                }
-            </div>
-        );
+            );
+        }
     }
 }
